requirements.txt
django-ckeditor==5.6.1
django-heroku==0.3.1
gunicorn==19.9.0
